requirements.txt

# Machine Learning and Data Science (for RAG)
scikit-learn>=1.4.0
numpy>=1.26.0

# Data manipulation and analysis
pandas>=2.1.0

# Additional utilities
requests>=2.31.0
pydantic>=2.0.0

# For better async handling in notebooks/streamlit
nest-asyncio>=1.6.0

# Environment management
python-dotenv>=1.0.0

# BERT and Transformers (for fine-tuning)
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
peft>=0.6.0
accelerate>=0.20.0
bitsandbytes>=0.41.0

# Jupyter and visualization
jupyter>=1.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Additional ML utilities
psutil>=5.9.0
google-generativeai>=0.3.0
shap>=0.44.0
newsapi-python>=0.2.7
newspaper3k>=0.2.8

# Web scraping and parsing
lxml>=4.9.0
lxml_html_clean>=0.1.0

# Hyperparameter tuning
optuna>=3.0.0

# Async HTTP requests
aiohttp>=3.8.0

# Streamlit for web app
streamlit>=1.28.0

# HuggingFace XET (for model storage)
hf_xet

# SentencePiece tokenizer (for some transformers)
sentencepiece>=0.1.99